requirements.txt
flask
intasend-python
requests

